requirements.txt
# FastAPI and Web Framework
fastapi==0.104.1
uvicorn==0.24.0
gunicorn==21.2.0
python-multipart==0.0.6
jinja2==3.1.2

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
asyncpg==0.29.0

# Authentication & OAuth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
authlib==1.2.1

# Background Tasks
celery==5.3.4
flower==2.0.1
kombu==5.3.4

# API Integrations
google-api-python-client==2.108.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
microsoft-graph-auth==0.2.0
microsoftgraph-python==1.0.0
slack-sdk==3.23.0
PyGithub==1.59.1
notion-client==2.2.1
aiohttp==3.9.1

# Email & Notifications
emails==0.6.0
pydantic-settings==2.0.3
smtplib-ssl==1.0.1

# System Monitoring & Performance
prometheus-client==0.19.0
structlog==23.2.0
psutil==5.9.6
py-cpuinfo==9.0.0

# Templates & Formatting
jinja2==3.1.2
markdown==3.5.1
beautifulsoup4==4.12.2

# Utilities
pydantic==2.5.0
python-dateutil==2.8.2
pytz==2023.3
schedule==1.2.0
click==8.1.7
python-dotenv==1.0.0

# Database Extensions
sqlalchemy-utils==0.41.1
databases[postgresql]==0.8.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
isort==5.13.0
bandit==1.7.5

# CORS & Middleware
fastapi-cors==0.0.6
starlette==0.27.0

# Security & Encryption
cryptography==41.0.8
bcrypt==4.1.2
PyJWT==2.8.0

# File Handling & Compression
pathlib==1.0.1
gzip-reader==0.1.0
zipfile36==0.1.3

# HTTP Clients & Networking
requests==2.31.0
urllib3==2.1.0
certifi==2023.11.17

# Async Extensions
aiofiles==23.2.1
aiosqlite==0.19.0
aiodns==3.1.1

# Configuration Management
python-config==0.1.2
configparser==6.0.0

# Data Validation & Serialization
marshmallow==3.20.2
jsonschema==4.20.0

# Caching
python-memcached==1.62
diskcache==5.6.3

# Logging Extensions
colorlog==6.8.0
loguru==0.7.2

# Error Handling & Debugging
sentry-sdk[fastapi]==1.40.0
traceback2==1.4.0
